% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assigned_windows_drive_letter}
\alias{assigned_windows_drive_letter}
\title{Drive letters for Windows}
\usage{
assigned_windows_drive_letter()
}
\value{
Drive letters if OS is windows
}
\description{
Drive letters for Windows
}
